+++ src/containers/Form/FormLayout.test.helper.ts
@@ -2,7 +2,7 @@ import { GET_TAG, GET_TAGS, GET_TAGS_COUNT, FILTER_TAGS } from '../../graphql/qu
 import { GET_LANGUAGES } from '../../graphql/queries/List';
 import { CREATE_TAG, DELETE_TAG, UPDATE_TAG } from '../../graphql/mutations/Tag';
 import { Input } from '../../components/UI/Form/Input/Input';
-import { getTagsQuery } from '../Chat/ChatMessages/ChatMessages.test.helper';
+import { getTagsQuery } from '../../mocks/Tag';
 
 export const listItemProps = {
   deleteItemQuery: DELETE_TAG,
