+++ src/containers/Profile/Contact/ContactProfile.test.tsx
@@ -3,7 +3,7 @@ import { render, wait } from '@testing-library/react';
 import { MockedProvider } from '@apollo/client/testing';
 
 import { ContactProfile } from './ContactProfile';
-import { LOGGED_IN_USER_MOCK } from '../Profile.test.helper';
+import { LOGGED_IN_USER_MOCK } from '../../../mocks/Contact';
 
 const mocks = LOGGED_IN_USER_MOCK;
