+++ src/containers/Profile/User/UserProfile.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { MockedProvider } from '@apollo/client/testing';
 
 import { UserProfile } from './UserProfile';
-import { LOGGED_IN_USER_MOCK } from '../Profile.test.helper';
+import { LOGGED_IN_USER_MOCK } from '../../../mocks/Contact';
 
 const mocks = LOGGED_IN_USER_MOCK;
